requirements.txt
langchain~=0.3.23
langchain-groq
langchain_chroma
langchain_community
langgraph~=0.3.28
langchain_huggingface
sentence_transformers
bs4~=0.0.2

beautifulsoup4~=4.13.3
python-dotenv~=1.1.0
typing_extensions~=4.13.2